+++ component/frontend/View/SubscriptionStatistics/Json.php
@@ -1,0 +1,14 @@
+<?php
+/**
+ * @package   AkeebaSubs
+ * @copyright Copyright (c)2010-2020 Nicholas K. Dionysopoulos / Akeeba Ltd
+ * @license   GNU General Public License version 3, or later
+ */
+
+namespace Akeeba\Subscriptions\Site\View\SubscriptionStatistics;
+
+defined('_JEXEC') or die;
+
+class Json extends \Akeeba\Subscriptions\Admin\View\SubscriptionStatistics\Json
+{
+}